- experimental valve action filter bug fix
git-svn-id: https://openhr20.svn.sourceforge.net/svnroot/openhr20@80 e6282f77-c345-0410-9e5f-9d7f1dee61d8

+++ trunk/source/controller.c
@@ -140,7 +140,7 @@ uint8_t CTL_update(bool minute_ch, uint8_t valve) {
                 if ((new_valve==0) || (new_valve==100)) {
                     valve = new_valve;
                 } else {
-                    int8_t x = CTL_last_dir*(int8_t)(valve-new_valve);
+                    int8_t x = CTL_last_dir*((int8_t)valve-(int8_t)new_valve);
                     if ((x>20) || (x<=0))  {
                         valve = new_valve;
                     }
